Task fix
